+++ __test__/schema/arblistDecimalsTooHigh.tokenlist.json
