+++ __test__/schema/arblistDecimalsTooLow.tokenlist.json
