+++ __test__/schema/arblistSymbolTooLong.tokenlist.json
