+++ __test__/schema/uniswap.tokenlist.json
